requirements.txt
requests
Flask
pyopenssl
simple_salesforce
python-dotenv